app: update imports for package move
