+++ cmd/frontend/codyapp/update_check.go
@@ -1,5 +1,5 @@
 // TODO(burmudar): move this app update endpoint code to codyapp package
-package updatecheck
+package codyapp
 
 import (
 	"context"
